vc-api/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
fastapi==0.101.1
flatdict==4.0.1
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
